requirements.txt
openai
streamlit
Pillow
moviepy
pyttsx3
requests